domain/memory/identities/wallets/miners/blocks/blocks_test.go
package blocks

import (
	"testing"

	mined_blocks "github.com/xmn-services/buckets-network/domain/memory/blocks/mined"
)

func TestBlocks_withoutBlocks_Success(t *testing.T) {
	ins, err := NewFactory().Create()
	if err != nil {
		t.Errorf("the error was expected to nil, error returned: %s", err.Error())
		return
	}

	adapter := NewAdapter()
	jsIns := adapter.ToJSON(ins)
	retIns, err := adapter.ToBlocks(jsIns)
	if err != nil {
		t.Errorf("the error was expected to nil, error returned: %s", err.Error())
		return
	}

	TestCompare(t, ins, retIns)
}

func TestBlocks_addBlock_deleteBlock_Success(t *testing.T) {
	ins, err := NewFactory().Create()
	if err != nil {
		t.Errorf("the error was expected to nil, error returned: %s", err.Error())
		return
	}

	_, minedBlockIns := mined_blocks.CreateBlockForTestsWithoutParams()
	err = ins.Add(minedBlockIns)
	if err != nil {
		t.Errorf("the error was expected to nil, error returned: %s", err.Error())
		return
	}

	err = ins.Delete(minedBlockIns.Hash())
	if err != nil {
		t.Errorf("the error was expected to nil, error returned: %s", err.Error())
		return
	}

	adapter := NewAdapter()
	jsIns := adapter.ToJSON(ins)
	retIns, err := adapter.ToBlocks(jsIns)
	if err != nil {
		t.Errorf("the error was expected to nil, error returned: %s", err.Error())
		return
	}

	TestCompare(t, ins, retIns)
}

func TestBlocks_addBlock_Success(t *testing.T) {
	ins, err := NewFactory().Create()
	if err != nil {
		t.Errorf("the error was expected to nil, error returned: %s", err.Error())
		return
	}

	_, minedBlockIns := mined_blocks.CreateBlockForTestsWithoutParams()
	err = ins.Add(minedBlockIns)
	if err != nil {
		t.Errorf("the error was expected to nil, error returned: %s", err.Error())
		return
	}

	adapter := NewAdapter()
	jsIns := adapter.ToJSON(ins)
	retIns, err := adapter.ToBlocks(jsIns)
	if err != nil {
		t.Errorf("the error was expected to nil, error returned: %s", err.Error())
		return
	}

	TestCompare(t, ins, retIns)
}
